% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_api.R
\name{gw_add_batch}
\alias{gw_add_batch}
\title{City Address Batch API}
\usage{
gw_add_batch(.data, id, address, threshold, vars = "minimal", crs)
}
\arguments{
\item{.data}{Name of data.frame containing address and id variables}

\item{id}{Name of column with unique identifier}

\item{address}{Vector containing addresses (List or Data.frame column)}

\item{threshold}{Numeric from 1 to 100, specifying how precise returned matches should be}

\item{vars}{How many variables should be returned? Choices are \code{"minmal"}, \code{"moderate"},
or \code{"all"}.}

\item{crs}{Output spatial reference (Not yet implemented)}
}
\value{
Returns a data.frame with the API response
}
\description{
City Address Batch API
}
