% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_api.R
\name{gw_add_candidates}
\alias{gw_add_candidates}
\title{City Address Candidates API Access}
\usage{
gw_add_candidates(street, zip, address, n, threshold, crs, sf)
}
\arguments{
\item{street}{Name of street}

\item{zip}{5-digit zipcode}

\item{address}{Single line address}

\item{n}{Number of candidates to return}

\item{threshold}{Numeric from 1 to 100, specifying how precise returned matches should be}

\item{crs}{Output spatial reference (CRS, WKID, ESRI Code)}

\item{sf}{Logical, output as simple feature object}
}
\value{
A data.frame or sf containing candidate addresses, x and y coordinates, and score of match
}
\description{
Provides access to the City of Saint Louis\href{https://stlgis3.stlouis-mo.gov/arcgis/rest/services/PUBLIC/COMPPARSTRZIPHANDLE/GeocodeServer/findAddressCandidates}{Address Candidates API}
}
