% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_api.R
\name{gw_add_reverse}
\alias{gw_add_reverse}
\title{City Reverse Geocoder}
\usage{
gw_add_reverse(x, y, distance = 0, crs = 102696,
  intersection = FALSE)
}
\arguments{
\item{x}{Numeric X coordinate of the location to reverse geocode}

\item{y}{Numeric Y coordinate of the location to reverse geocode}

\item{distance}{The distance in meters from the given location within which a matching address should be searched. If this parameter is not provided or an invalid value is provided, a default value of 0 meters is used.}

\item{crs}{Numeric CRS or WKID for spatial projection}

\item{intersection}{Logical, Return nearest Address (FALSE) or Intersection (TRUE)}
}
\description{
City Reverse Geocoder
}
