% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_validation.R
\name{gw_address_validation}
\alias{gw_address_validation}
\title{City Address Validation API}
\usage{
gw_address_validation(search, key)
}
\arguments{
\item{search}{A string containing a single address or 11-digit parcel ID}

\item{key}{Your API key (will be stored in Renviron in future update)}
}
\value{
A data.frame with 59 variables for all addresses matching the search query
}
\description{
Provides access to the City of Saint Louis\href{https://www.stlouis-mo.gov/government/departments/information-technology/web-development/city-api/address-validation-api.cfm}{Address Validation API}
}
\details{
An API key is required to use this function, you can obtain one \href{https://www.stlouis-mo.gov/government/departments/information-technology/web-development/city-api/sign-up.cfm}{here}.
}
