% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_aggregate.R
\name{gw_aggregate_points}
\alias{gw_aggregate_points}
\title{Aggregate Points}
\usage{
gw_aggregate_points(.data, to, sf = TRUE, replace.na = TRUE,
  keep.na = FALSE)
}
\arguments{
\item{.data}{A \code{sf} object}

\item{to}{The string name of an areal unit to aggregate to: block group, tract,
precinct, ward, neighborhood, or municipality.}

\item{sf}{A logical scalar; if \code{TRUE}, returns an \code{sf} object. Otherwise returns
a tibble.}

\item{replace.na}{A logical scalar; if \code{TRUE}, areal units that do not have any points
enclosed in them with be given a value of \code{0}. If \code{FALSE}, they will be given
a value of \code{NA}.}

\item{keep.na}{A logical scalar; if \code{TRUE}, a row with count of points that could not be
joined to the areal unit will be returned. This occurs when points fall outside of all
given features. If \code{FALSE}, no count of missing points is returned. This argument
only returns an \code{NA} row if \code{sf = FALSE}.}
}
\description{
Performs a spatial join to obtain counts of points within specified
    areal units. This is done with a range of canned options representing
    common scenarios in cleaning spatial data.
}
