% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_validation.R
\name{gw_api_key}
\alias{gw_api_key}
\title{Install a City of St. Louis API Key in Your \code{.Renviron} File for Repeated Use}
\usage{
gw_api_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{Your quoted API key.}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing CENSUS_API_KEY
that you already have in your \code{.Renviron} file.}

\item{install}{A logical scalar; if \code{TRUE}, your key will be installed in your \code{.Renviron}
file for use in future sessions.  If \code{FALSE} (default), your key will be available only
for your current session.}
}
\description{
Install a City of St. Louis API Key in Your \code{.Renviron} File for Repeated Use
}
