% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{gw_build_geocoder}
\alias{gw_build_geocoder}
\title{Build Local Geocoder}
\usage{
gw_build_geocoder(style, crs = 4269, return = c("coords", "parcel", "zip"),
    include_units = FALSE, appendix)
}
\arguments{
\item{style}{One of either \code{"full"} (\code{"123 Main St"}) or \code{"short"} (\code{"123 Main"}).}

\item{crs}{A numeric code corresponding to the desired coordinate system for the column output if
\code{return} includes \code{"coords"} as well as the object output if \code{class} is \code{"sf"}.}

\item{return}{Optional; A character scalar or vector that describes the type of information to be applied
to the target data. Options include the City's address identification numbers (\code{addrrecnum}),
parcel identification numbers (\code{handle}), zip-codes, and x and y coordinates (in decimal degrees).}

\item{include_units}{A logical scalar; if \code{TRUE}, all individual records for apartment units will
be included. If \code{FALSE} (default), only records for the overall building will be retained.}

\item{appendix}{An object with additional address or placename data to append to the geocoder.}
}
\value{
A tibble with the requested data in the \code{return} argument as well as
   a \code{address} variable containing the full street address string.
}
\description{
This constructs a local geocoder that is based on the latest release of the City of
   St. Louis's master address list. This function therefore requires an internet connection,
   and will error if your computer is offline. Since the actual geocoding is done with a second
   function, however, it is possible to build a geocoder and store it offline for repeated use.
}
\seealso{
\code{\link{gw_geocode}}, \href{https://www.stlouis-mo.gov/data/geocode-service.cfm}{City of St. Louis Geocode Service}
}
