% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{gw_geocode}
\alias{gw_geocode}
\title{Geocode Addresses}
\usage{
gw_geocode(.data, type, var, zip, class, local, local_short, local_place, threshold)
}
\arguments{
\item{.data}{A target data set}

\item{type}{Geocoder type; one of either \code{"local"}, \code{"local short"}, \code{"city batch"},
\code{"city candidate"}, \code{"census"}, or \code{"osm"}. Composite geocoders consisting of
only local geocoders (\code{"local"}, \code{"local short"}, and \code{"local placename"}) or
both local and the remote geocoders are also available by using \code{"composite, local"} or
\code{"composite, full"} respectively.}

\item{var}{Address variable in the target data set, which should contain the house number,
street directionals, name, and suffix.}

\item{zip}{Name of zipcode variable in the taget data set (optional).}

\item{class}{Output class; one of either \code{"sf"} or \code{"tibble"}.}

\item{local}{Name of object containing a local geocoder built with
\code{\link{gw_build_geocoder}}}

\item{local_short}{Name of object containing a local, short style geocoder
built with \code{\link{gw_build_geocoder}}}

\item{local_place}{Name of object containing a local, placename style geocoder
built with \code{\link{gw_build_geocoder}}}

\item{threshold}{For the city candidate geocoder, what score is the minimum acceptable?}
}
\value{
A copy of the target data with georeferenced data applied to it.
}
\description{
Apply a previously build geocoder to target data. This function will
   apply whatever unique variables exist in the geocoder. See \code{\link{gw_build_geocoder}}
   for options.
}
\seealso{
\code{\link{gw_build_geocoder}}
}
