% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames.R
\name{gw_geonames}
\alias{gw_geonames}
\title{Access to the geonames general search}
\usage{
gw_geonames(search, username, max_row = 10, fuzzy = 1, country = "US")
}
\arguments{
\item{search}{A string containing search query}

\item{username}{A string with your geonames username}

\item{max_row}{Single integer specifying how many rows you would like returned at maximum}

\item{fuzzy}{Numeric from > 0 to 1 specifying how strict to match spelling}

\item{country}{String containing country abbreviation to limit results}
}
\value{
A data.frame containing the geonames response, if no results, returns an empty list
}
\description{
Requires a geonames username. Register for a geonames.org username at http://www.geonames.org/login/ and then you will need to enable access to the geonames webservices by going to http://www.geonames.org/enablefreewebservice
}
