% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_get_city.R
\name{gw_get_city}
\alias{gw_get_city}
\title{Get Data from the City of St. Louis}
\usage{
gw_get_city(data)
}
\arguments{
\item{data}{A character vector matching the name of a data source available on the City of
St. Louis's open data website.}
}
\value{
\code{gw_get_city} returns a table or simple features object with the requested data.
}
\description{
The \code{gw_get_city}...
}
