% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{gw_get_coords}
\alias{gw_get_coords}
\title{Extract Coordinates from sf Object}
\usage{
gw_get_coords(.data, names = c("x","y"), crs = 4269)
}
\arguments{
\item{.data}{A \code{sf} object}

\item{names}{A vector with two column names, one for the x coordinate and one for the y coordinate.}

\item{crs}{A numeric code corresponding to the desired coordinate system for the column output}
}
\value{
An updated object with two new columns based on the names provided in the \code{names} argument.
}
\description{
Converts point coordinates stored in an \code{sf} object to columns for
    both the x and y coordinates. Useful for storing spatial data in tabular form.
}
\details{
Based on a function written \href{https://github.com/jmlondon}{Josh M. London} and
    described in a \href{https://github.com/r-spatial/sf/issues/231}{GitHub issue}.
}
