% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{gw_get_data}
\alias{gw_get_data}
\title{Download Data About St. Louis}
\usage{
gw_get_data(data, class, ...)
}
\arguments{
\item{data}{A character string describing the data you wish to download.}

\item{class}{Either \code{"sf"} to return a simple features object or code
\code{"tibble"} to return a tibble.}

\item{...}{Optional arguments to pass on to \pkg{tidycensus} or \pkg{tigris}}
}
\value{
A \code{sf} object or tibble with the requested data.
}
\description{
Download data from the City of St. Louis's open data page or the Saint Louis
    University openGIS project's repositories.
}
\details{
Data provided by the City of St. Louis are provided without warrenty. Data from the
    Saint Louis openGIS project are provided under a
    \href{https://opendatacommons.org/licenses/by/index.html}{Open Data Commons Attribution License}.

The following data are currently available:

\describe{
  \item{\code{Addresses}}{City of St. Louis; street address master list (\code{sf} or tibble)}
  \item{\code{Grids}}{SLU OpenGIS; square kilometer grids for spatial statistics (\code{sf})}
  \item{\code{Grids, Clipped}}{SLU OpenGIS; square kilometer grids for spatial statistics with major parks clipped out (\code{sf})}
  \item{\code{Grids, Exploded}}{SLU OpenGIS; square kilometer grids for spatial statistics with major parks clipped out and slivers exploded (\code{sf})}
  \item{\code{Land Records}}{City of St. Louis; City property records (tibble)}
  \item{\code{Land Use}}{City of St. Louis; land use data (\code{sf} or tibble)}
  \item{\code{Mississippi River, Illinois}}{SLU OpenGIS; extent of the Mississippi River in Madison and St. Clair counties for the length of the City of St. Louis (\code{sf})}
  \item{\code{Mississippi River, Islands}}{SLU OpenGIS; group of islands on the Illinois side of the Mississippi (\code{sf})}
  \item{\code{Neighborhoods}}{City of St. Louis; City neighborhoods (\code{sf} or tibble)}
  \item{\code{Ortho Tile Boundaries}}{SLU OpenGIS; Eest-West Gateway Orthoimagery Tile Boundaries (\code{sf} or tibble)}
  \item{\code{Parcels}}{City of St. Louis; parcel data (\code{sf} or tibble)}
  \item{\code{Parks}}{City of St. Louis; public parks (\code{sf} or tibble)}
  \item{\code{Police Districts}}{City of St. Louis; police districts (\code{sf} or tibble)}
  \item{\code{Police Districts, Pre-2014}}{City of St. Louis; police districts prior to 2014
      reorganization (\code{sf} or tibble)}
  \item{\code{Tracts}}{\pkg{tigris}; Census Tracts for the City of St. Louis (\code{sf} or tibble)}
  \item{\code{Voting Precincts}}{City of St. Louis; current voting precincts (\code{sf} or tibble)}
  \item{\code{Zoning}}{City of St. Louis; zoning data (\code{sf} or tibble)}
  \item{\code{Zoning, Multi}}{City of St. Louis; zoning data (\code{sf} or tibble)}
}
}
