% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_aggregate.R
\name{gw_get_epsg}
\alias{gw_get_epsg}
\title{Extract epsg value}
\usage{
gw_get_epsg(.data)
}
\arguments{
\item{.data}{A \code{sf} object}
}
\description{
Returns the numeric epsg value from \code{sf::st_crs}.
}
