% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_get_repo.R
\name{gw_get_repo}
\alias{gw_get_repo}
\title{Download openGIS Data Repositories}
\usage{
gw_get_repo(repo)
}
\arguments{
\item{repo}{A character vector matching the name of a repository to download and extract
data from}
}
\value{
\code{gw_get_repo} returns a simple features object with the requested data.
}
\description{
\code{gw_get_repo} provides access to shapefiles and \code{.geoJSON} files stored on the
St. Louis openGIS GitHub organization.
}
