% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{gw_identify}
\alias{gw_identify}
\title{Add Areal Unit Identifiers to Point Data}
\usage{
gw_identify(.data, to, side = "right", class = "sf")
}
\arguments{
\item{.data}{A \code{sf} object}

\item{to}{The string name of an areal unit to aggregate to: \code{"block group"}, \code{"tract"},
\code{"precinct"}, \code{"ward"}, or \code{"neighborhood"}.}

\item{side}{One of either \code{"right"} or \code{"left"} indicating where the identifier variable
should be placed in the}

\item{class}{One of either \code{"sf"} (default) or \code{"tibble"}.}
}
\value{
\code{gw_identify} returns either a \code{sf} object or tibble with the requested data.
}
\description{
Apply identifiers from a selection of commonly used areal units
   to point data stored in a \code{sf} object.
}
