% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_aggregate.R
\name{gw_is_sf}
\alias{gw_is_sf}
\title{Logic check for sf data}
\usage{
gw_is_sf(.data)
}
\arguments{
\item{.data}{An object to be tested}
}
\description{
Retuns a \code{TRUE} or \code{FALSE} value based on whether
    the object is from class \code{sf} or not.
}
