% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_aggregate.R
\name{gw_load_areal}
\alias{gw_load_areal}
\title{Load areal data}
\usage{
gw_load_areal(name)
}
\arguments{
\item{name}{The name of the areal geometry to be loaded}
}
\description{
Based on arguments supplied, areal data is loaded, re-projected,
    and has its ID varaible renamed to a generic \code{ID} to facilitate
    simpler code in \code{gw_aggregate_city} or \code{gw_aggregate_county}.
}
