% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{gw_nhood}
\alias{gw_nhood}
\title{Neighborhood Indentifier Conversion}
\usage{
gw_nhood(.data, var, new_var, from, to)
}
\arguments{
\item{.data}{Data frame, tibble, or \code{sf} object to be modified}

\item{var}{Variable to base conversions on}

\item{new_var}{Name of new variable; if none is specified, the \code{var} variable
will be overwritten}

\item{from}{The output format of the data \code{var} (one of either \code{"numeric"} or
\code{"string"}).}
}
\description{
Provides a technique for converting neighborhood identifiers between
    their numeric, string, and factor formats.
}
