% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_parseAddress.R
\name{gw_parseAddress}
\alias{gw_parseAddress}
\title{Parse Street Addresses}
\usage{
gw_parseAddress(.data, address, houseNum = TRUE, overwrite = TRUE,
    keepVars = TRUE, locale = 29510)
}
\arguments{
\item{.data}{A tbl}

\item{address}{A character vector with street address information
formatted as either "123 Main St" or "Main St"}

\item{houseNum}{A logical scalar. Are house numbers included in
the \code{address} vector?}

\item{overwrite}{A logical scalar. Should the output overwrite
the given variable?}

\item{keepVars}{A logical scalar. Should the specific address
components be returned as individual variables?}

\item{locale}{A numeric scalar. The only current accepted
value is 29510 for the City of St. Louis. This will apply specific
street name corrections that are applicable to St. Louis.}
}
\value{
\code{gw_stName} returns a tibble with the requested output -
either the original data with the street address overwritten in the
given vector or the original data with a set of new individual variables
representing specific address components.
}
\description{
\code{gw_parseAddress} is designed to process a single vector that contains
a house number and street information including prefix direction, street
name, and street type. The inclusion of the house number is optional.
These data look like either "123 Main St" or "Main St". With the
\code{keepVars} argument, users can decide whether they want to simply
return a cleaned version of their input with standardized spelling and
formatting, or whether they also want to return individual variables
representing specific address components.
}
\note{
\code{gw_parseAddress} requires that a number variable names be unused
in the original data - \code{houseNum}, \code{houseNumL}, \code{houseNumU},
\code{stFull}, \code{stDir}, \code{stName}, \code{stType}. If these names
are present, \code{gw_parseAddress} will return an error.
}
