% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_aggregate.R
\name{gw_rename_id}
\alias{gw_rename_id}
\title{Rename areal ID variable}
\usage{
gw_rename_id(.data, name)
}
\arguments{
\item{.data}{A \code{sf} object}

\item{name}{The name of the areal geometry that has been loaded}
}
\description{
Rename the ID variable after processing is done so that it
    matches the origianl ID variable in the data.
}
