% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_stName.R
\name{gw_stName}
\alias{gw_stName}
\title{Standardize Street Names}
\usage{
gw_stName(data, variable, overwrite = TRUE, newVariable)
}
\arguments{
\item{data}{A tibble or a data frame}

\item{variable}{A character vector within \code{data} that contains City of St. Louis street names}

\item{overwrite}{A logical scalar. Should the output overwrite the given variable?}

\item{newVariable}{A name for a new vector to be created if \code{overwrite = FALSE}}
}
\value{
\code{gw_stName} returns a tibble with the requested output - either the original data with
non-standard street names overwritten in the given vector or the original data with a new
vector containing standardized street names.
}
\description{
\code{gw_stName} standardizes a given set of City of St. Louis street names by checking
them against a master list of street names and returning a preferred name. For example,
"1ST" and "FIRST" are both possible variations of "FIRST ST". If "FIRST" is given,
\code{gw_stdName} will offer "1ST" as the standardized alternative.
}
\note{
\code{gw_stName} requires two variable names be unused in the original data - \code{dc_corrrect} and
\code{dc_incorrect}. If these names are present, \code{gw_stName} will return an error.
}
