% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhood_Str.R
\name{nhood_Str}
\alias{nhood_Str}
\title{String or factor neighborhood names}
\source{
\href{https://www.stlouis-mo.gov/government/departments/planning/documents/citywide-neighborhood-map.cfm}{City of St. Louis Citywide Neighborhood Map}
}
\usage{
nhood_Str(var, asFactor = TRUE)
}
\arguments{
\item{var}{A numeric or integer vector containing City of St. Louis neighborhood numbers}

\item{asFactor}{A logical scalar. Should the output be converted to a factor?}
}
\description{
\code{nhood_Str} returns the matching neighborhood name in string or factor format
}
\details{
This is a function for converting a vector of City of St. Louis neighborhood numbers
to either string or factor format.
}
\examples{
if (requireNamespace("dplyr", quietly = TRUE)) {

    neighborhoods <- data.frame(
        id = c(1, 2, 3, 4, 5),
        nhoodNum = c(1, 2, 4, 5, 3)
    )

    neighborhoods <- dplyr::mutate(neighborhoods, nhood = nhood_Str(nhoodNum, asFactor = TRUE))
    table(neighborhoods$nhood)
}

}
