% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmastr.R
\docType{data}
\name{stl_std_directions}
\alias{stl_std_directions}
\title{Standardized Directions}
\format{A data frame with 7 rows and 2 variables:
\describe{
  \item{dir.output}{directional abbreviation}
  \item{dir.input}{directional abbreviation}
  }}
\usage{
stl_std_directions
}
\description{
A data set containing standardized street directionals for St. Louis.
    This can be used as the basis for a dictionary appendix for the \pkg{postmastr}
    package. It contains only the extra directionals used in the City's master list
    of street addresses.
}
\examples{
head(stl_std_directions)

}
\keyword{datasets}
