% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmastr.R
\docType{data}
\name{stl_std_houseSuffix}
\alias{stl_std_houseSuffix}
\title{Standardized House Suffix Values}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{houseSuf.output}{standardized abbreviation for house suffix}
  \item{houseSuf.input}{alternative house suffix}
  }
}
\usage{
stl_std_houseSuffix
}
\description{
A data set containing standardized house suffix names and possible
    alternatives. This can be used as the basis for a dictionary appendix for
    the \pkg{postmastr} package.
}
\examples{
head(stl_std_houseSuffix)

}
\keyword{datasets}
