% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_std_streets.R
\docType{data}
\name{stl_std_streets}
\alias{stl_std_streets}
\title{Standardized Street Names}
\format{A data frame with 12 rows and 2 variables:
\describe{
  \item{correct}{standardized street name}
  \item{incorrect}{alternative street name}
  }}
\usage{
stl_std_streets
}
\description{
A data set containing standardized street names and possible alternatives.
    This can be used as the basis for a dictionary appendix for the \pkg{postmastr}
    package.
}
\examples{
head(stl_std_streets)

}
\keyword{datasets}
