% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmastr.R
\docType{data}
\name{stl_std_suffix}
\alias{stl_std_suffix}
\title{Standardized Street Suffix Names}
\format{A data frame with 76 rows and 3 variables:
\describe{
  \item{suf.type}{suffix type}
  \item{suf.input}{input suffix}
  \item{suf.output}{output suffix}
  }}
\usage{
stl_std_suffix
}
\description{
A data set containing the street suffixes used in St. Louis and possible
    alternatives. This can be used as the basis for a dictionary appendix for
    the \pkg{postmastr} package.
}
\examples{
head(stl_std_suffix)

}
\keyword{datasets}
