% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_tracts10.R
\docType{data}
\name{stl_tracts10}
\alias{stl_tracts10}
\title{City of St. Louis Census Tracts, 2010}
\format{A tibble with 106 rows and 5 variables:
\describe{
  \item{GEOID}{identification number for feature}
  \item{TRACTCE}{census tract number}
  \item{ALAND}{land area in meters^2}
  \item{AWATER}{water area in meters^2}
  \item{geometry}{\code{sf} object geometric data}
  }}
\source{
U.S. Census Bureau via \code{tigris} package
}
\usage{
data(stl_tracts10)
}
\description{
An \code{sf} object that contains a polygon for each 2010 tract in St. Louis.
}
\examples{
str(stl_tracts10)
head(stl_tracts10)

}
\keyword{datasets}
