% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_wards10.R
\docType{data}
\name{stl_wards10}
\alias{stl_wards10}
\title{City of St. Louis Wards, 2010}
\format{A tibble with 28 rows and 3 variables:
\describe{
  \item{ID}{ward number}
  \item{AREA}{area in meters^2}
  \item{geometry}{\code{sf} object geometric data}
  }}
\source{
City of St. Louis
}
\usage{
data(stl_wards10)
}
\description{
An \code{sf} object that contains a polygon for each 2010 ward in St. Louis.
}
\examples{
str(stl_wards10)
head(stl_wards10)

}
\keyword{datasets}
