% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlCity.R
\docType{data}
\name{stlCity}
\alias{stlCity}
\title{City of St. Louis Boundary}
\format{A tibble with 1 rows and 5 variables:
\describe{
  \item{GEOID}{identification number for feature}
  \item{NAMELSAD}{string name for feature}
  \item{ALAND}{land area in meters^2}
  \item{AWATER}{water area in meters^2}
  \item{geometry}{\code{sf} object geometric data}
  }}
\source{
U.S. Census Bureau via \code{tigris} package
}
\usage{
data(stlCity)
}
\description{
An \code{sf} object that contains a single polygon for the entire city.
}
\examples{
str(stlCity)
head(stlCity)

}
\keyword{datasets}
