% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sushi.R
\docType{data}
\name{sushi_sf}
\alias{sushi_sf}
\title{Sushi Resturants in the Ciy of St. Louis, sf Format}
\format{A \code{sf} object with 15 rows and 4 variables:
\describe{
  \item{name}{resturant name}
  \item{address}{street address}
  \item{visit}{visit date}
  \item{geometry}{sf geometric data}
}}
\usage{
data(sushi_sf)
}
\description{
A selection of sushi resturants in the St. Louis, Missouri region.
   These data are in long form, with multiple visits in a single year to some
   of the resturants. These data are the parsed and standardized form of the
   \code{sushi2} data set in the \pkg{postmastr} package. They have been
   geocoded and projected, and so can be used to demonstate the aggregate
   and identify features of the \pkg{gateway} package.
}
\examples{
head(sushi_sf)

}
\keyword{datasets}
