% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{gw_build_geocoder}
\alias{gw_build_geocoder}
\title{Build Local Geocoder}
\usage{
gw_build_geocoder(return = c("coords", "parcel", "zip"), class,
  crs = 4269, include_units = FALSE)
}
\arguments{
\item{return}{A character scalar or vector that describes the type of information to be applied
to the target data. Options include the City's address identification numbers (\code{addrrecnum}),
parcel identification numbers (\code{handle}), zip-codes, and x and y coordinates (in decimal degrees).}

\item{class}{One of either \code{"sf"} or \code{"tibble"}.}

\item{include_units}{A logical scalar; if \code{TRUE}, all individual records for apartment units will
be included. If \code{FALSE} (default), only records for the overall building will be retained.}
}
\value{
A \code{sf} object or tibble with the requested data in the \code{return} argument as well as
   a \code{address} variable containing the full street address string.
}
\description{
This constructs a local geocoder that is based on the latest release of the City of
   St. Louis's master address list. This function therefore requires an internet connection,
   and will error if your computer is offline. Since the actual geocoding is done with a second
   function, however, it is possible to build a geocoder and store it offline for repeated use.
}
\seealso{
\code{\link{gw_geocode}}, \href{https://www.stlouis-mo.gov/data/geocode-service.cfm}{City of St. Louis Geocode Service}
}
