% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{gw_geocode}
\alias{gw_geocode}
\title{Geocode Addresses}
\usage{
gw_geocode(.data, type, class, address, geocoder, include_result = TRUE)
}
\arguments{
\item{.data}{A target data set}

\item{type}{Geocoder type; one of either \code{"local"}, \code{"city api"}, or \code{"census"}.}

\item{class}{Output class; one of either \code{"sf"} or \code{"tibble"}.}

\item{address}{Address variable in the target data set, which should contain the house number,
street directionals, name, and suffix, and optionally unit types and numbers as well. Unit
names should be replaced with \code{#} to match how \code{\link{gw_build_geocoder}}
creates units.}

\item{geocoder}{Name of object containing a geocoder built with \code{\link{gw_build_geocoder}}}

\item{include_result}{Logical scalar; if \code{TRUE} (default), a column describing how each
observation was geocoded is included in the output.}
}
\value{
A copy of the target data with georeferenced data applied to it.
}
\description{
Apply a previously build geocoder to target data. This function will
   apply whatever unique variables exist in the geocoder. See \code{\link{gw_build_geocoder}}
   for options.
}
\seealso{
\code{\link{gw_build_geocoder}}
}
