% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{gw_geocode_composite}
\alias{gw_geocode_composite}
\title{Composite Geocoder}
\usage{
gw_geocode_composite(.data, var, local_geocoder, short_geocoder)
}
\arguments{
\item{.data}{A data frame or tibble to be geocoded}

\item{var}{Column with address data to be geocoded}

\item{local_geocoder}{Object with local geocoder data}

\item{short_geocoder}{Object with short version of local geocoder data}
}
\description{
An algorithm for processing and geocoding address data. A first
    attempt is made to match again a local geocoder. Unmatched addresses are then
    matched against a short geocoder. AAddress that remain unmatched are then
    matched using the City of St. Louis's address candidate API.
}
