% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhood_num.R
\name{nhood_num}
\alias{nhood_num}
\title{Numeric neighborhood numbers}
\source{
\href{https://www.stlouis-mo.gov/government/departments/planning/documents/citywide-neighborhood-map.cfm}{City of St. Louis Citywide Neighborhood Map}
}
\usage{
nhood_num(var)
}
\arguments{
\item{var}{A string vector containing City of St. Louis neighborhood names}
}
\description{
\code{nhood_num} returns the matching neighborhood name in string or factor format
}
\details{
This is a function for converting a vector of City of St. Louis neighborhood numbers
to either string or factor format.
}
\note{
Function is capable of handeling vectors that contain various spellings (i.e. "carondelet" or "Carondelet").
It is also capable of handleing some common neighborhood abbreviations, such as "Bevo", "CWE", "FPSE", and "The Grove".
}
\examples{
if (requireNamespace("dplyr", quietly = TRUE)) {

    neighborhoods <- data.frame(
        id = c(1, 2, 3, 4, 5),
        nhoodStr = c("Patch", "Bevo", "Bevo Mill", "Lindenwood Park", "Carondelet")
    )

    neighborhoods <- dplyr::mutate(neighborhoods, nhood = nhood_num(nhoodStr))
    table(neighborhoods$nhood)
}

}
