% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdStreets.R
\docType{data}
\name{stdStreets}
\alias{stdStreets}
\title{Standardized Street Names}
\format{A data frame with 106 rows and 7 variables:
\describe{
  \item{dc_correct}{standardized street name}
  \item{dc_incorrect}{alternative street name}
  }}
\usage{
data(stdStreets)
}
\description{
A data set containing standardized street names and possible alternatives.
}
\keyword{datasets}
