% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdSuffix.R
\name{stdSuffix}
\alias{stdSuffix}
\title{Standardized Street Suffixes}
\usage{
stdSuffix(.data, variable, overwrite = TRUE, newVariable)
}
\arguments{
\item{.data}{A tibble or a data frame}

\item{variable}{A character vector within \code{.data} that contains City of St. Louis street suffixes}

\item{overwrite}{A logical scalar. Should the output overwrite the given variable?}

\item{newVariable}{A name for a new vector to be created if \code{overwrite = FALSE}}
}
\value{
\code{stdSuffix} returns a tibble with the requested output - either the suffix variable has
    been overwritten or a new variable with corrected suffix abbreviations has been added.
}
\description{
\code{stdSuffix} standardizes a given set of street suffix values to the USPS preferred suffix abbreviation.
}
\note{
\code{stdSuffix} requires two variable names be unused in the original data - \code{suf_com} and
\code{suf_pri}. If these names are present, \code{stdSuffix} will return an error.
}
