% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sushi.R
\docType{data}
\name{sushi}
\alias{sushi}
\title{Sushi Resturants in the Ciy of St. Louis, Tibble Format}
\format{
A tibble with 15 rows and 3 variables:
\describe{
  \item{name}{resturant name}
  \item{address}{street address}
  \item{visit}{visit date}
}
}
\usage{
data(sushi)
}
\description{
A selection of sushi resturants in the St. Louis, Missouri region.
   These data are in long form, with multiple visits in a single year to some
   of the resturants. These data are the parsed and standardized form of the
   \code{sushi2} data set in the \pkg{postmastr} package. They are therefore
   useful for demonstrating the geocoding features of the \pkg{gateway} package.
}
\examples{
head(sushi)

}
\keyword{datasets}
